enable support for visible annotations on mashups
